Extract player dial page component
